+++ apps/commerce-api/src/main/java/com/loopers/application/like/event/LikeEventHandler.java
@@ -1,17 +1,11 @@
 package com.loopers.application.like.event;
 
 import com.loopers.application.like.LikeProcessingService;
-import com.loopers.domain.like.LikeType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.orm.ObjectOptimisticLockingFailureException;
-import org.springframework.retry.annotation.Backoff;
-import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Component;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
 
 /**
  * 좋아요 관련 이벤트 처리 핸들러
