+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderFacade.java
@@ -49,8 +49,6 @@ public class OrderFacade {
     private final ApplicationEventPublisher eventPublisher;
     private final KafkaEventPublisher kafkaEventPublisher;
 
-    private final ApplicationEventPublisher eventPublisher;
-
     @Transactional
     public Order placeOrder(String userId, OrderInfo orderInfo) {
