--- apps/commerce-collector/src/main/java/com/loopers/collector/infrastructure/eventhandled/EventHandledRepositoryImpl.java
@@ -1,25 +1,0 @@
-package com.loopers.collector.infrastructure.eventhandled;
-
-import com.loopers.collector.domain.eventhandled.EventHandled;
-import com.loopers.collector.domain.eventhandled.EventHandledRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class EventHandledRepositoryImpl implements EventHandledRepository {
-
-    private final EventHandledJpaRepository eventHandledJpaRepository;
-
-    @Override
-    public boolean existsByEventId(String eventId) {
-        return eventHandledJpaRepository.existsByEventId(eventId);
-    }
-
-    // INSERT 또는 UPDATE 자동 판단
-    @Override
-    public EventHandled save(EventHandled eventHandled) {
-        return eventHandledJpaRepository.save(eventHandled);
-    }
-
-}
