--- apps/commerce-collector/src/main/java/com/loopers/consumer/EventEnvelope.java
@@ -1,40 +1,0 @@
-package com.loopers.consumer;
-
-import java.time.Instant;
-
-/**
- * Kafka 이벤트 수신을 위한 EventEnvelope
- *
- * 목적:
- * - Producer(commerce-api)에서 전송한 EventEnvelope 구조와 동일
- * - Consumer(commerce-collector)에서 이벤트 메타데이터와 payload 분리 처리
- *
- * 구조 설명:
- * - eventType: 이벤트 타입 식별 (StockDecreasedEvent, LikeAddedEvent 등)
- * - eventId: Producer에서 생성한 고유 ID (멱등성 보장용)
- * - timestamp: Producer에서 이벤트 발행한 시간
- * - payload: 실제 이벤트 데이터 (StockDecreasedEvent 객체 등)
- *
- * JSON 예시:
- * {
- *   "eventType": "StockDecreasedEvent",
- *   "eventId": "evt_abc123456",
- *   "timestamp": "2024-01-01T10:00:00Z",
- *   "payload": {
- *     "productId": 123,
- *     "previousStock": 10,
- *     "currentStock": 8,
- *     "decreasedQuantity": 2,
- *     "reason": "ORDER",
- *     "occurredAt": "2024-01-01T10:00:00Z"
- *   }
- * }
- */
-public record EventEnvelope(
-    String eventType,     // 이벤트 타입 (Consumer 타입 식별용)
-    String eventId,       // 고유 ID (멱등성 처리용, Producer가 생성)
-    Instant timestamp,    // 이벤트 발행 시간 (Producer에서 설정)
-    Object payload        // 실제 이벤트 데이터 (JSON 역직렬화 후 Object)
-) {
-
-}
